Iniital commit This is a conversion of the previous example page to TypeScript and to use the NPM package for types. It works all the way up to the packaging, which I was forced to use as browsers cannot import without the .js extension, and the types package imports are not compiled by TypeScript so no way to change that without a build tool.
Sadly the build tool is not working for me yet, not sure what to do about that, but many of the things are ready.

Left to do are all the output messages, they are not using the types yet so will automatically fail as they will not parse to the classes in OpenVR2WS. That will be fixed whenever I can get the build tool to work.